Filter out agentless global data tags from input add_fields processor
