+++ x-pack/platform/plugins/shared/fleet/server/services/agent_policies/package_policies_to_agent_inputs.ts
@@ -8,6 +8,7 @@ import { merge } from 'lodash';
 import deepMerge from 'deepmerge';
 
 import type { FullAgentPolicyAddFields, GlobalDataTag } from '../../../common/types';
+import { getAgentlessGlobalDataTags } from '../../../common/services/agentless_policy_helper';
 import { isPackageLimited } from '../../../common/services';
 import type {
   PackagePolicy,
@@ -155,11 +156,6 @@ export const storedPackagePoliciesToAgentInputs = async (
 ): Promise<FullAgentPolicyInput[]> => {
   const fullInputs: FullAgentPolicyInput[] = [];
 
-  const addFields =
-    globalDataTags && globalDataTags.length > 0
-      ? globalDataTagsToAddFields(globalDataTags)
-      : undefined;
-
   for (const packagePolicy of packagePolicies) {
     if (!isPolicyEnabled(packagePolicy)) {
       continue;
@@ -168,6 +164,12 @@ export const storedPackagePoliciesToAgentInputs = async (
     const packageInfo = packagePolicy.package
       ? packageInfoCache.get(pkgToPkgKey(packagePolicy.package))
       : undefined;
+
+    const filteredGlobalDataTags = filterGlobalDataTags(globalDataTags, packageInfo);
+    const addFields =
+      filteredGlobalDataTags && filteredGlobalDataTags.length > 0
+        ? globalDataTagsToAddFields(filteredGlobalDataTags)
+        : undefined;
 
     fullInputs.push(
       ...storedPackagePolicyToAgentInputs(
@@ -196,4 +198,21 @@ const globalDataTagsToAddFields = (tags: GlobalDataTag[]): FullAgentPolicyAddFie
       fields,
     },
   };
+};
+
+const filterGlobalDataTags = (
+  globalDataTags: GlobalDataTag[] | undefined,
+  packageInfo: PackageInfo | undefined
+): GlobalDataTag[] | undefined => {
+  const agentlessGlobalDataTags = getAgentlessGlobalDataTags(packageInfo);
+
+  if (!globalDataTags || !packageInfo || !agentlessGlobalDataTags) {
+    return globalDataTags;
+  }
+
+  return globalDataTags.filter((globalDataTag) => {
+    return !agentlessGlobalDataTags.some(
+      ({ name, value }) => name === globalDataTag.name && value === globalDataTag.value
+    );
+  });
 };
